Add score radar chart in Detail Page

+++ app/src/main/java/com/jim/moviecritics/data/Score.kt
@@ -7,10 +7,10 @@ data class Score(
     val userID: Long,
     val imdbID: String,
     val createdTime: Timestamp,
-    val leisure: Double,
-    val hit: Double,
-    val cast: Double,
-    val music: Double,
-    val story: Double,
-    val average: Double
+    val leisure: Float,
+    val hit: Float,
+    val cast: Float,
+    val music: Float,
+    val story: Float,
+    val average: Float
 )
